Refactor all logic into dedicated files for each metric and refactor css/test files
