+++ src/Counters/Pooing.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useCountUp } from 'react-countup';
-import {currentNumberOfShaggers} from '../counter-logic/counter-logic'
-import {addNoise} from '../counter-logic/counter-logic'
+import {currentNumberOfShaggers} from '../counter-logic/SexLogic'
+import {addNoise} from '../counter-logic/SexLogic'
 
 
 const PooingCounter = () => {
@@ -18,7 +18,7 @@ const PooingCounter = () => {
 
     function updateCounter(){
         // update called twice below so there is a nice transition from initial count up and noise
-        update((currentNumberOfShaggers() * 10) + addNoise() -1)
+        update((currentNumberOfShaggers() * 10) + addNoise())
 
         const timer = setInterval(() => {
             update((currentNumberOfShaggers() * 10) + addNoise())
