+++ src/counter-logic/SexLogic.js
@@ -32,4 +32,3 @@ export function addNoise(maxFluctuation){
 
     return randomNormal({mean: 0, dev: deviation})
 }
-
